archive/training/requirements-dev.txt
# Development and training dependencies
# Install with: pip install -r requirements.txt -r requirements-dev.txt

# Training and evaluation
scikit-learn==1.3.0
seaborn==0.12.2
pandas==2.0.3
tqdm==4.66.1
torchmetrics==1.1.2

# Development tools
requests  # for test_api.py

# Jupyter (if needed for experimentation)
# jupyter
# matplotlib
